Remove CAS1 Assessments API Generated code
As part of this change `crnOrName` was removed from the `getAssessmentsForUser` API because it is not used
